+++ Controller/AdminController.php
@@ -11,739 +11,9 @@
 
 namespace JavierEguiluz\Bundle\EasyAdminBundle\Controller;
 
-use Doctrine\DBAL\Exception\ForeignKeyConstraintViolationException;
-use Doctrine\ORM\EntityManager;
-use Doctrine\ORM\QueryBuilder;
-use JavierEguiluz\Bundle\EasyAdminBundle\Event\EasyAdminEvents;
-use JavierEguiluz\Bundle\EasyAdminBundle\Exception\EntityRemoveException;
-use JavierEguiluz\Bundle\EasyAdminBundle\Exception\ForbiddenActionException;
-use JavierEguiluz\Bundle\EasyAdminBundle\Exception\NoEntitiesConfiguredException;
-use JavierEguiluz\Bundle\EasyAdminBundle\Exception\UndefinedEntityException;
-use JavierEguiluz\Bundle\EasyAdminBundle\Form\Util\LegacyFormHelper;
-use Pagerfanta\Pagerfanta;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-use Symfony\Component\EventDispatcher\GenericEvent;
-use Symfony\Component\Form\Form;
-use Symfony\Component\Form\FormBuilder;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\Form\FormInterface;
-use Symfony\Component\HttpFoundation\JsonResponse;
-use Symfony\Component\HttpFoundation\RedirectResponse;
-use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpFoundation\Response;
-
 /**
- * The controller used to render all the default EasyAdmin actions.
- *
- * @author Javier Eguiluz <javier.eguiluz@gmail.com>
+ * @deprecated use \JavierEguiluz\Bundle\EasyAdminBundle\Controller\EasyAdminController instead.
  */
-class AdminController extends Controller
+class AdminController extends EasyAdminController
 {
-    /** @var array The full configuration of the entire backend */
-    protected $config;
-    /** @var array The full configuration of the current entity */
-    protected $entity = array();
-    /** @var Request The instance of the current Symfony request */
-    protected $request;
-    /** @var EntityManager The Doctrine entity manager for the current entity */
-    protected $em;
-
-    /**
-     * @Route("/", name="easyadmin")
-     * @Route("/", name="admin")
-     *
-     * The 'admin' route is deprecated since version 1.8.0 and it will be removed in 2.0.
-     *
-     * @param Request $request
-     *
-     * @return RedirectResponse|Response
-     */
-    public function indexAction(Request $request)
-    {
-        $this->initialize($request);
-
-        if (null === $request->query->get('entity')) {
-            return $this->redirectToBackendHomepage();
-        }
-
-        $action = $request->query->get('action', 'list');
-        if (!$this->isActionAllowed($action)) {
-            throw new ForbiddenActionException(array('action' => $action, 'entity_name' => $this->entity['name']));
-        }
-
-        return $this->executeDynamicMethod($action.'<EntityName>Action');
-    }
-
-    /**
-     * Utility method which initializes the configuration of the entity on which
-     * the user is performing the action.
-     *
-     * @param Request $request
-     */
-    protected function initialize(Request $request)
-    {
-        $this->dispatch(EasyAdminEvents::PRE_INITIALIZE);
-
-        $this->config = $this->get('easyadmin.config.manager')->getBackendConfig();
-
-        if (0 === count($this->config['entities'])) {
-            throw new NoEntitiesConfiguredException();
-        }
-
-        // this condition happens when accessing the backend homepage and before
-        // redirecting to the default page set as the homepage
-        if (null === $entityName = $request->query->get('entity')) {
-            return;
-        }
-
-        if (!array_key_exists($entityName, $this->config['entities'])) {
-            throw new UndefinedEntityException(array('entity_name' => $entityName));
-        }
-
-        $this->entity = $this->get('easyadmin.config.manager')->getEntityConfiguration($entityName);
-
-        $action = $request->query->get('action', 'list');
-        if (!$request->query->has('sortField')) {
-            $sortField = isset($this->entity[$action]['sort']['field']) ? $this->entity[$action]['sort']['field'] : $this->entity['primary_key_field_name'];
-            $request->query->set('sortField', $sortField);
-        }
-        if (!$request->query->has('sortDirection')) {
-            $sortDirection = isset($this->entity[$action]['sort']['direction']) ? $this->entity[$action]['sort']['direction'] : 'DESC';
-            $request->query->set('sortDirection', $sortDirection);
-        }
-
-        $this->em = $this->getDoctrine()->getManagerForClass($this->entity['class']);
-        $this->request = $request;
-
-        $this->dispatch(EasyAdminEvents::POST_INITIALIZE);
-    }
-
-    protected function dispatch($eventName, array $arguments = array())
-    {
-        $arguments = array_replace(array(
-            'config' => $this->config,
-            'em' => $this->em,
-            'entity' => $this->entity,
-            'request' => $this->request,
-        ), $arguments);
-
-        $subject = isset($arguments['paginator']) ? $arguments['paginator'] : $arguments['entity'];
-        $event = new GenericEvent($subject, $arguments);
-
-        $this->get('event_dispatcher')->dispatch($eventName, $event);
-    }
-
-    /**
-     * The method that returns the values displayed by an autocomplete field
-     * based on the user's input.
-     *
-     * @return JsonResponse
-     */
-    protected function autocompleteAction()
-    {
-        $results = $this->get('easyadmin.autocomplete')->find(
-            $this->request->query->get('entity'),
-            $this->request->query->get('query'),
-            $this->request->query->get('page', 1)
-        );
-
-        return new JsonResponse($results);
-    }
-
-    /**
-     * The method that is executed when the user performs a 'list' action on an entity.
-     *
-     * @return Response
-     */
-    protected function listAction()
-    {
-        $this->dispatch(EasyAdminEvents::PRE_LIST);
-
-        $fields = $this->entity['list']['fields'];
-        $paginator = $this->findAll($this->entity['class'], $this->request->query->get('page', 1), $this->config['list']['max_results'], $this->request->query->get('sortField'), $this->request->query->get('sortDirection'), $this->entity['list']['dql_filter']);
-
-        $this->dispatch(EasyAdminEvents::POST_LIST, array('paginator' => $paginator));
-
-        return $this->render($this->entity['templates']['list'], array(
-            'paginator' => $paginator,
-            'fields' => $fields,
-            'delete_form_template' => $this->createDeleteForm($this->entity['name'], '__id__')->createView(),
-        ));
-    }
-
-    /**
-     * The method that is executed when the user performs a 'edit' action on an entity.
-     *
-     * @return Response|RedirectResponse
-     */
-    protected function editAction()
-    {
-        $this->dispatch(EasyAdminEvents::PRE_EDIT);
-
-        $id = $this->request->query->get('id');
-        $easyadmin = $this->request->attributes->get('easyadmin');
-        $entity = $easyadmin['item'];
-
-        if ($this->request->isXmlHttpRequest() && $property = $this->request->query->get('property')) {
-            $newValue = 'true' === mb_strtolower($this->request->query->get('newValue'));
-            $fieldsMetadata = $this->entity['list']['fields'];
-
-            if (!isset($fieldsMetadata[$property]) || 'toggle' !== $fieldsMetadata[$property]['dataType']) {
-                throw new \RuntimeException(sprintf('The type of the "%s" property is not "toggle".', $property));
-            }
-
-            $this->updateEntityProperty($entity, $property, $newValue);
-
-            return new Response((string) $newValue);
-        }
-
-        $fields = $this->entity['edit']['fields'];
-
-        $editForm = $this->executeDynamicMethod('create<EntityName>EditForm', array($entity, $fields));
-        $deleteForm = $this->createDeleteForm($this->entity['name'], $id);
-
-        $editForm->handleRequest($this->request);
-        if ($editForm->isSubmitted() && $editForm->isValid()) {
-            $this->dispatch(EasyAdminEvents::PRE_UPDATE, array('entity' => $entity));
-
-            $this->executeDynamicMethod('preUpdate<EntityName>Entity', array($entity));
-            $this->em->flush();
-
-            $this->dispatch(EasyAdminEvents::POST_UPDATE, array('entity' => $entity));
-
-            $refererUrl = $this->request->query->get('referer', '');
-
-            return !empty($refererUrl)
-                ? $this->redirect(urldecode($refererUrl))
-                : $this->redirect($this->generateUrl('easyadmin', array('action' => 'list', 'entity' => $this->entity['name'])));
-        }
-
-        $this->dispatch(EasyAdminEvents::POST_EDIT);
-
-        return $this->render($this->entity['templates']['edit'], array(
-            'form' => $editForm->createView(),
-            'entity_fields' => $fields,
-            'entity' => $entity,
-            'delete_form' => $deleteForm->createView(),
-        ));
-    }
-
-    /**
-     * The method that is executed when the user performs a 'show' action on an entity.
-     *
-     * @return Response
-     */
-    protected function showAction()
-    {
-        $this->dispatch(EasyAdminEvents::PRE_SHOW);
-
-        $id = $this->request->query->get('id');
-        $easyadmin = $this->request->attributes->get('easyadmin');
-        $entity = $easyadmin['item'];
-
-        $fields = $this->entity['show']['fields'];
-        $deleteForm = $this->createDeleteForm($this->entity['name'], $id);
-
-        $this->dispatch(EasyAdminEvents::POST_SHOW, array(
-            'deleteForm' => $deleteForm,
-            'fields' => $fields,
-            'entity' => $entity,
-        ));
-
-        return $this->render($this->entity['templates']['show'], array(
-            'entity' => $entity,
-            'fields' => $fields,
-            'delete_form' => $deleteForm->createView(),
-        ));
-    }
-
-    /**
-     * The method that is executed when the user performs a 'new' action on an entity.
-     *
-     * @return Response|RedirectResponse
-     */
-    protected function newAction()
-    {
-        $this->dispatch(EasyAdminEvents::PRE_NEW);
-
-        $entity = $this->executeDynamicMethod('createNew<EntityName>Entity');
-
-        $easyadmin = $this->request->attributes->get('easyadmin');
-        $easyadmin['item'] = $entity;
-        $this->request->attributes->set('easyadmin', $easyadmin);
-
-        $fields = $this->entity['new']['fields'];
-
-        $newForm = $this->executeDynamicMethod('create<EntityName>NewForm', array($entity, $fields));
-
-        $newForm->handleRequest($this->request);
-        if ($newForm->isSubmitted() && $newForm->isValid()) {
-            $this->dispatch(EasyAdminEvents::PRE_PERSIST, array('entity' => $entity));
-
-            $this->executeDynamicMethod('prePersist<EntityName>Entity', array($entity));
-
-            $this->em->persist($entity);
-            $this->em->flush();
-
-            $this->dispatch(EasyAdminEvents::POST_PERSIST, array('entity' => $entity));
-
-            $refererUrl = $this->request->query->get('referer', '');
-
-            return !empty($refererUrl)
-                ? $this->redirect(urldecode($refererUrl))
-                : $this->redirect($this->generateUrl('easyadmin', array('action' => 'list', 'entity' => $this->entity['name'])));
-        }
-
-        $this->dispatch(EasyAdminEvents::POST_NEW, array(
-            'entity_fields' => $fields,
-            'form' => $newForm,
-            'entity' => $entity,
-        ));
-
-        return $this->render($this->entity['templates']['new'], array(
-            'form' => $newForm->createView(),
-            'entity_fields' => $fields,
-            'entity' => $entity,
-        ));
-    }
-
-    /**
-     * The method that is executed when the user performs a 'delete' action to
-     * remove any entity.
-     *
-     * @return RedirectResponse
-     */
-    protected function deleteAction()
-    {
-        $this->dispatch(EasyAdminEvents::PRE_DELETE);
-
-        if ('DELETE' !== $this->request->getMethod()) {
-            return $this->redirect($this->generateUrl('easyadmin', array('action' => 'list', 'entity' => $this->entity['name'])));
-        }
-
-        $id = $this->request->query->get('id');
-        $form = $this->createDeleteForm($this->entity['name'], $id);
-        $form->handleRequest($this->request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $easyadmin = $this->request->attributes->get('easyadmin');
-            $entity = $easyadmin['item'];
-
-            $this->dispatch(EasyAdminEvents::PRE_REMOVE, array('entity' => $entity));
-
-            $this->executeDynamicMethod('preRemove<EntityName>Entity', array($entity));
-
-            try {
-                $this->em->remove($entity);
-                $this->em->flush();
-            } catch (ForeignKeyConstraintViolationException $e) {
-                throw new EntityRemoveException(array('entity_name' => $this->entity['name']));
-            }
-
-            $this->dispatch(EasyAdminEvents::POST_REMOVE, array('entity' => $entity));
-        }
-
-        $refererUrl = $this->request->query->get('referer', '');
-
-        $this->dispatch(EasyAdminEvents::POST_DELETE);
-
-        return !empty($refererUrl)
-            ? $this->redirect(urldecode($refererUrl))
-            : $this->redirect($this->generateUrl('easyadmin', array('action' => 'list', 'entity' => $this->entity['name'])));
-    }
-
-    /**
-     * The method that is executed when the user performs a query on an entity.
-     *
-     * @return Response
-     */
-    protected function searchAction()
-    {
-        $this->dispatch(EasyAdminEvents::PRE_SEARCH);
-
-        // if the search query is empty, redirect to the 'list' action
-        if ('' === $this->request->query->get('query')) {
-            $queryParameters = array_replace($this->request->query->all(), array('action' => 'list', 'query' => null));
-            $queryParameters = array_filter($queryParameters);
-
-            return $this->redirect($this->get('router')->generate('easyadmin', $queryParameters));
-        }
-
-        $searchableFields = $this->entity['search']['fields'];
-        $paginator = $this->findBy($this->entity['class'], $this->request->query->get('query'), $searchableFields, $this->request->query->get('page', 1), $this->config['list']['max_results'], $this->request->query->get('sortField'), $this->request->query->get('sortDirection'), $this->entity['search']['dql_filter']);
-        $fields = $this->entity['list']['fields'];
-
-        $this->dispatch(EasyAdminEvents::POST_SEARCH, array(
-            'fields' => $fields,
-            'paginator' => $paginator,
-        ));
-
-        return $this->render($this->entity['templates']['list'], array(
-            'paginator' => $paginator,
-            'fields' => $fields,
-            'delete_form_template' => $this->createDeleteForm($this->entity['name'], '__id__')->createView(),
-        ));
-    }
-
-    /**
-     * It updates the value of some property of some entity to the new given value.
-     *
-     * @param mixed  $entity   The instance of the entity to modify
-     * @param string $property The name of the property to change
-     * @param bool   $value    The new value of the property
-     *
-     * @throws \RuntimeException
-     */
-    protected function updateEntityProperty($entity, $property, $value)
-    {
-        $entityConfig = $this->entity;
-
-        // the method_exists() check is needed because Symfony 2.3 doesn't have isWritable() method
-        if (method_exists($this->get('property_accessor'), 'isWritable') && !$this->get('property_accessor')->isWritable($entity, $property)) {
-            throw new \RuntimeException(sprintf('The "%s" property of the "%s" entity is not writable.', $property, $entityConfig['name']));
-        }
-
-        $this->dispatch(EasyAdminEvents::PRE_UPDATE, array('entity' => $entity, 'newValue' => $value));
-
-        $this->get('property_accessor')->setValue($entity, $property, $value);
-
-        $this->em->persist($entity);
-        $this->em->flush();
-        $this->dispatch(EasyAdminEvents::POST_UPDATE, array('entity' => $entity, 'newValue' => $value));
-
-        $this->dispatch(EasyAdminEvents::POST_EDIT);
-    }
-
-    /**
-     * Creates a new object of the current managed entity.
-     * This method is mostly here for override convenience, because it allows
-     * the user to use his own method to customize the entity instantiation.
-     *
-     * @return object
-     */
-    protected function createNewEntity()
-    {
-        $entityFullyQualifiedClassName = $this->entity['class'];
-
-        return new $entityFullyQualifiedClassName();
-    }
-
-    /**
-     * Allows applications to modify the entity associated with the item being
-     * created before persisting it.
-     *
-     * @param object $entity
-     */
-    protected function prePersistEntity($entity)
-    {
-    }
-
-    /**
-     * Allows applications to modify the entity associated with the item being
-     * edited before persisting it.
-     *
-     * @param object $entity
-     */
-    protected function preUpdateEntity($entity)
-    {
-    }
-
-    /**
-     * Allows applications to modify the entity associated with the item being
-     * deleted before removing it.
-     *
-     * @param object $entity
-     */
-    protected function preRemoveEntity($entity)
-    {
-    }
-
-    /**
-     * Performs a database query to get all the records related to the given
-     * entity. It supports pagination and field sorting.
-     *
-     * @param string      $entityClass
-     * @param int         $page
-     * @param int         $maxPerPage
-     * @param string|null $sortField
-     * @param string|null $sortDirection
-     * @param string|null $dqlFilter
-     *
-     * @return Pagerfanta The paginated query results
-     */
-    protected function findAll($entityClass, $page = 1, $maxPerPage = 15, $sortField = null, $sortDirection = null, $dqlFilter = null)
-    {
-        if (empty($sortDirection) || !in_array(strtoupper($sortDirection), array('ASC', 'DESC'))) {
-            $sortDirection = 'DESC';
-        }
-
-        $queryBuilder = $this->executeDynamicMethod('create<EntityName>ListQueryBuilder', array($entityClass, $sortDirection, $sortField, $dqlFilter));
-
-        $this->dispatch(EasyAdminEvents::POST_LIST_QUERY_BUILDER, array(
-            'query_builder' => $queryBuilder,
-            'sort_field' => $sortField,
-            'sort_direction' => $sortDirection,
-        ));
-
-        return $this->get('easyadmin.paginator')->createOrmPaginator($queryBuilder, $page, $maxPerPage);
-    }
-
-    /**
-     * Creates Query Builder instance for all the records.
-     *
-     * @param string      $entityClass
-     * @param string      $sortDirection
-     * @param string|null $sortField
-     * @param string|null $dqlFilter
-     *
-     * @return QueryBuilder The Query Builder instance
-     */
-    protected function createListQueryBuilder($entityClass, $sortDirection, $sortField = null, $dqlFilter = null)
-    {
-        return $this->get('easyadmin.query_builder')->createListQueryBuilder($this->entity, $sortField, $sortDirection, $dqlFilter);
-    }
-
-    /**
-     * Performs a database query based on the search query provided by the user.
-     * It supports pagination and field sorting.
-     *
-     * @param string      $entityClass
-     * @param string      $searchQuery
-     * @param array       $searchableFields
-     * @param int         $page
-     * @param int         $maxPerPage
-     * @param string|null $sortField
-     * @param string|null $sortDirection
-     * @param string|null $dqlFilter
-     *
-     * @return Pagerfanta The paginated query results
-     */
-    protected function findBy($entityClass, $searchQuery, array $searchableFields, $page = 1, $maxPerPage = 15, $sortField = null, $sortDirection = null, $dqlFilter = null)
-    {
-        $queryBuilder = $this->executeDynamicMethod('create<EntityName>SearchQueryBuilder', array($entityClass, $searchQuery, $searchableFields, $sortField, $sortDirection, $dqlFilter));
-
-        $this->dispatch(EasyAdminEvents::POST_SEARCH_QUERY_BUILDER, array(
-            'query_builder' => $queryBuilder,
-            'search_query' => $searchQuery,
-            'searchable_fields' => $searchableFields,
-        ));
-
-        return $this->get('easyadmin.paginator')->createOrmPaginator($queryBuilder, $page, $maxPerPage);
-    }
-
-    /**
-     * Creates Query Builder instance for search query.
-     *
-     * @param string      $entityClass
-     * @param string      $searchQuery
-     * @param array       $searchableFields
-     * @param string|null $sortField
-     * @param string|null $sortDirection
-     * @param string|null $dqlFilter
-     *
-     * @return QueryBuilder The Query Builder instance
-     */
-    protected function createSearchQueryBuilder($entityClass, $searchQuery, array $searchableFields, $sortField = null, $sortDirection = null, $dqlFilter = null)
-    {
-        return $this->get('easyadmin.query_builder')->createSearchQueryBuilder($this->entity, $searchQuery, $sortField, $sortDirection, $dqlFilter);
-    }
-
-    /**
-     * Creates the form used to edit an entity.
-     *
-     * @param object $entity
-     * @param array  $entityProperties
-     *
-     * @return Form
-     */
-    protected function createEditForm($entity, array $entityProperties)
-    {
-        return $this->createEntityForm($entity, $entityProperties, 'edit');
-    }
-
-    /**
-     * Creates the form used to create an entity.
-     *
-     * @param object $entity
-     * @param array  $entityProperties
-     *
-     * @return Form
-     */
-    protected function createNewForm($entity, array $entityProperties)
-    {
-        return $this->createEntityForm($entity, $entityProperties, 'new');
-    }
-
-    /**
-     * Creates the form builder of the form used to create or edit the given entity.
-     *
-     * @param object $entity
-     * @param string $view   The name of the view where this form is used ('new' or 'edit')
-     *
-     * @return FormBuilder
-     */
-    protected function createEntityFormBuilder($entity, $view)
-    {
-        $formOptions = $this->executeDynamicMethod('get<EntityName>EntityFormOptions', array($entity, $view));
-
-        return $this->get('form.factory')->createNamedBuilder(mb_strtolower($this->entity['name']), LegacyFormHelper::getType('easyadmin'), $entity, $formOptions);
-    }
-
-    /**
-     * Retrieves the list of form options before sending them to the form builder.
-     * This allows adding dynamic logic to the default form options.
-     *
-     * @param object $entity
-     * @param string $view
-     *
-     * @return array
-     */
-    protected function getEntityFormOptions($entity, $view)
-    {
-        $formOptions = $this->entity[$view]['form_options'];
-        $formOptions['entity'] = $this->entity['name'];
-        $formOptions['view'] = $view;
-
-        return $formOptions;
-    }
-
-    /**
-     * Creates the form object used to create or edit the given entity.
-     *
-     * @param object $entity
-     * @param array  $entityProperties
-     * @param string $view
-     *
-     * @return FormInterface
-     *
-     * @throws \Exception
-     */
-    protected function createEntityForm($entity, array $entityProperties, $view)
-    {
-        if (method_exists($this, $customMethodName = 'create'.$this->entity['name'].'EntityForm')) {
-            $form = $this->{$customMethodName}($entity, $entityProperties, $view);
-            if (!$form instanceof FormInterface) {
-                throw new \UnexpectedValueException(sprintf(
-                    'The "%s" method must return a FormInterface, "%s" given.',
-                    $customMethodName, is_object($form) ? get_class($form) : gettype($form)
-                ));
-            }
-
-            return $form;
-        }
-
-        $formBuilder = $this->executeDynamicMethod('create<EntityName>EntityFormBuilder', array($entity, $view));
-
-        if (!$formBuilder instanceof FormBuilderInterface) {
-            throw new \UnexpectedValueException(sprintf(
-                'The "%s" method must return a FormBuilderInterface, "%s" given.',
-                'createEntityForm', is_object($formBuilder) ? get_class($formBuilder) : gettype($formBuilder)
-            ));
-        }
-
-        return $formBuilder->getForm();
-    }
-
-    /**
-     * Creates the form used to delete an entity. It must be a form because
-     * the deletion of the entity are always performed with the 'DELETE' HTTP method,
-     * which requires a form to work in the current browsers.
-     *
-     * @param string     $entityName
-     * @param int|string $entityId   When reusing the delete form for multiple entities, a pattern string is passed instead of an integer
-     *
-     * @return Form|FormInterface
-     */
-    protected function createDeleteForm($entityName, $entityId)
-    {
-        /** @var FormBuilder $formBuilder */
-        $formBuilder = $this->get('form.factory')->createNamedBuilder('delete_form')
-            ->setAction($this->generateUrl('easyadmin', array('action' => 'delete', 'entity' => $entityName, 'id' => $entityId)))
-            ->setMethod('DELETE')
-        ;
-        $formBuilder->add('submit', LegacyFormHelper::getType('submit'), array('label' => 'delete_modal.action', 'translation_domain' => 'EasyAdminBundle'));
-        // needed to avoid submitting empty delete forms (see issue #1409)
-        $formBuilder->add('_easyadmin_delete_flag', LegacyFormHelper::getType('hidden'), array('data' => '1'));
-
-        return $formBuilder->getForm();
-    }
-
-    /**
-     * Utility method that checks if the given action is allowed for
-     * the current entity.
-     *
-     * @param string $actionName
-     *
-     * @return bool
-     */
-    protected function isActionAllowed($actionName)
-    {
-        return false === in_array($actionName, $this->entity['disabled_actions'], true);
-    }
-
-    /**
-     * Utility shortcut to render an error when the requested action is not allowed
-     * for the given entity.
-     *
-     * @param string $action
-     *
-     * @deprecated Use the ForbiddenException instead of this method
-     *
-     * @return Response
-     */
-    protected function renderForbiddenActionError($action)
-    {
-        return $this->render('@EasyAdmin/error/forbidden_action.html.twig', array('action' => $action), new Response('', 403));
-    }
-
-    /**
-     * Given a method name pattern, it looks for the customized version of that
-     * method (based on the entity name) and executes it. If the custom method
-     * does not exist, it executes the regular method.
-     *
-     * For example:
-     *   executeDynamicMethod('create<EntityName>Entity') and the entity name is 'User'
-     *   if 'createUserEntity()' exists, execute it; otherwise execute 'createEntity()'
-     *
-     * @param string $methodNamePattern The pattern of the method name (dynamic parts are enclosed with <> angle brackets)
-     * @param array  $arguments         The arguments passed to the executed method
-     *
-     * @return mixed
-     */
-    protected function executeDynamicMethod($methodNamePattern, array $arguments = array())
-    {
-        $methodName = str_replace('<EntityName>', $this->entity['name'], $methodNamePattern);
-
-        if (!is_callable(array($this, $methodName))) {
-            $methodName = str_replace('<EntityName>', '', $methodNamePattern);
-        }
-
-        return call_user_func_array(array($this, $methodName), $arguments);
-    }
-
-    /**
-     * Generates the backend homepage and redirects to it.
-     */
-    protected function redirectToBackendHomepage()
-    {
-        $homepageConfig = $this->config['homepage'];
-
-        $url = isset($homepageConfig['url'])
-            ? $homepageConfig['url']
-            : $this->get('router')->generate($homepageConfig['route'], $homepageConfig['params']);
-
-        return $this->redirect($url);
-    }
-
-    /**
-     * It renders the main CSS applied to the backend design. This controller
-     * allows to generate dynamic CSS files that use variables without the need
-     * to set up a CSS preprocessing toolchain.
-     *
-     * @deprecated The CSS styles are no longer rendered at runtime but preprocessed during container compilation. Use the $container['easyadmin.config']['_internal']['custom_css'] variable instead
-     */
-    public function renderCssAction()
-    {
-    }
 }
